First set of doc updates, update versions to next.5, update to latest monaco-vscode-api 3.1.2, use monaco-editor as dependency designator
- Moved some code back from the wrapper to the languageclient as otherwise it can be used standalone
- Cleaned enhanced MonacoEnvironment and fixed init bug in new common init function
