+++ package.json
@@ -3,14 +3,14 @@
   "type": "module",
   "devDependencies": {
     "@codingame/esbuild-import-meta-url-plugin": "~1.0.2",
-    "@types/node": "~20.11.27",
-    "@types/react": "~18.2.65",
+    "@types/node": "~20.11.28",
+    "@types/react": "~18.2.66",
     "@types/react-dom": "~18.2.22",
     "@types/shelljs": "~0.8.15",
     "@types/vscode": "~1.87.0",
     "@typescript-eslint/eslint-plugin": "~7.2.0",
     "@typescript-eslint/parser": "~7.2.0",
-    "@vitest/browser": "~1.3.1",
+    "@vitest/browser": "~1.4.0",
     "editorconfig": "~2.0.0",
     "eslint": "~8.57.0",
     "eslint-plugin-header": "~3.1.1",
@@ -21,7 +21,7 @@
     "shx": "~0.3.4",
     "typescript": "~5.4.2",
     "vite": "~5.1.6",
-    "vitest": "~1.3.1",
+    "vitest": "~1.4.0",
     "webdriverio": "~8.33.1"
   },
   "volta": {
