+++ packages/examples/package.json
@@ -1,6 +1,6 @@
 {
   "name": "monaco-languageclient-examples",
-  "version": "8.0.0-next.4",
+  "version": "8.0.0-next.5",
   "description": "Monaco Language client examples",
   "author": {
     "name": "TypeFox GmbH",
@@ -54,39 +54,38 @@
     }
   },
   "dependencies": {
-    "@codingame/monaco-vscode-api": "~3.1.1",
-    "@codingame/monaco-vscode-editor-api": "~3.1.1",
-    "@codingame/monaco-vscode-configuration-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-editor-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-files-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-groovy-default-extension": "~3.1.1",
-    "@codingame/monaco-vscode-json-default-extension": "~3.1.1",
-    "@codingame/monaco-vscode-keybindings-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-python-default-extension": "~3.1.1",
-    "@codingame/monaco-vscode-textmate-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-theme-defaults-default-extension": "~3.1.1",
-    "@codingame/monaco-vscode-theme-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-javascript-default-extension": "~3.1.1",
-    "@codingame/monaco-vscode-typescript-basics-default-extension": "~3.1.1",
-    "@codingame/monaco-vscode-typescript-language-features-default-extension": "~3.1.1",
-    "@codingame/monaco-vscode-language-pack-de": "~3.1.1",
-    "@codingame/monaco-vscode-language-pack-es": "~3.1.1",
-    "@codingame/monaco-vscode-language-pack-fr": "~3.1.1",
-    "@typefox/monaco-editor-react": "~3.0.0-next.4",
+    "@codingame/monaco-vscode-configuration-service-override": "~3.1.2",
+    "@codingame/monaco-vscode-editor-service-override": "~3.1.2",
+    "@codingame/monaco-vscode-files-service-override": "~3.1.2",
+    "@codingame/monaco-vscode-groovy-default-extension": "~3.1.2",
+    "@codingame/monaco-vscode-json-default-extension": "~3.1.2",
+    "@codingame/monaco-vscode-keybindings-service-override": "~3.1.2",
+    "@codingame/monaco-vscode-python-default-extension": "~3.1.2",
+    "@codingame/monaco-vscode-textmate-service-override": "~3.1.2",
+    "@codingame/monaco-vscode-theme-defaults-default-extension": "~3.1.2",
+    "@codingame/monaco-vscode-theme-service-override": "~3.1.2",
+    "@codingame/monaco-vscode-javascript-default-extension": "~3.1.2",
+    "@codingame/monaco-vscode-typescript-basics-default-extension": "~3.1.2",
+    "@codingame/monaco-vscode-typescript-language-features-default-extension": "~3.1.2",
+    "@codingame/monaco-vscode-language-pack-de": "~3.1.2",
+    "@codingame/monaco-vscode-language-pack-es": "~3.1.2",
+    "@codingame/monaco-vscode-language-pack-fr": "~3.1.2",
+    "@typefox/monaco-editor-react": "~3.0.0-next.5",
     "express": "~4.18.3",
     "langium": "3.0.0",
-    "monaco-languageclient": "~8.0.0-next.4",
-    "monaco-editor-wrapper": "~4.0.0-next.4",
+    "monaco-editor": "npm:@codingame/monaco-vscode-editor-api@~3.1.2",
+    "monaco-editor-wrapper": "~4.0.0-next.5",
+    "monaco-languageclient": "~8.0.0-next.5",
     "pyright": "~1.1.354",
     "react": "~18.2.0",
     "react-dom": "~18.2.0",
     "request-light": "~0.7.0",
-    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.1",
+    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.2",
     "vscode-json-languageservice": "~5.3.7",
     "vscode-languageclient": "~9.0.1",
     "vscode-languageserver": "~9.0.1",
     "vscode-uri": "~3.0.8",
-    "vscode-ws-jsonrpc": "~3.3.0-next.4",
+    "vscode-ws-jsonrpc": "~3.3.0-next.5",
     "ws": "~8.16.0"
   },
   "devDependencies": {
