+++ packages/vscode-ws-jsonrpc/package.json
@@ -1,6 +1,6 @@
 {
   "name": "vscode-ws-jsonrpc",
-  "version": "3.3.0-next.4",
+  "version": "3.3.0-next.5",
   "description": "VSCode JSON RPC over WebSocket",
   "author": {
     "name": "TypeFox GmbH",
