+++ packages/wrapper-react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@typefox/monaco-editor-react",
-  "version": "3.0.0-next.4",
+  "version": "3.0.0-next.5",
   "license": "MIT",
   "description": "React component for Monaco-Editor and Monaco Languageclient",
   "keywords": [
@@ -50,14 +50,13 @@
     "npm": "10.2.4"
   },
   "dependencies": {
-    "@codingame/monaco-vscode-api": "~3.1.1",
-    "@codingame/monaco-vscode-editor-api": "~3.1.1",
-    "monaco-editor-wrapper": "~4.0.0-next.4",
+    "monaco-editor": "npm:@codingame/monaco-vscode-editor-api@~3.1.2",
+    "monaco-editor-wrapper": "~4.0.0-next.5",
     "react": "~18.2.0",
-    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.1"
+    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.2"
   },
   "peerDependencies": {
-    "monaco-editor-wrapper": "~4.0.0-next.4"
+    "monaco-editor-wrapper": "~4.0.0-next.5"
   },
   "peerDependenciesMeta": {
     "monaco-editor-wrapper": {
