+++ packages/wrapper/package.json
@@ -1,6 +1,6 @@
 {
   "name": "monaco-editor-wrapper",
-  "version": "4.0.0-next.4",
+  "version": "4.0.0-next.5",
   "license": "MIT",
   "description": "Wrapper for monaco-vscode-editor-api and monaco-languageclient",
   "keywords": [
@@ -101,34 +101,28 @@
     "npm": "10.2.4"
   },
   "dependencies": {
-    "@codingame/monaco-vscode-configuration-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-editor-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-textmate-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-theme-defaults-default-extension": "~3.1.1",
-    "@codingame/monaco-vscode-theme-service-override": "~3.1.1",
-    "@codingame/monaco-vscode-api": "~3.1.1",
-    "@codingame/monaco-vscode-editor-api": "~3.1.1",
-    "esbuild": "~0.20.1",
-    "monaco-languageclient": "~8.0.0-next.4",
-    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.1",
+    "@codingame/monaco-vscode-configuration-service-override": "~3.1.2",
+    "@codingame/monaco-vscode-editor-service-override": "~3.1.2",
+    "@codingame/monaco-vscode-textmate-service-override": "~3.1.2",
+    "@codingame/monaco-vscode-theme-defaults-default-extension": "~3.1.2",
+    "@codingame/monaco-vscode-theme-service-override": "~3.1.2",
+    "esbuild": "~0.20.2",
+    "monaco-editor": "npm:@codingame/monaco-vscode-editor-api@~3.1.2",
+    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.2",
     "vscode-languageclient": "~9.0.1",
     "vscode-languageserver-protocol": "~3.17.5",
-    "vscode-ws-jsonrpc": "~3.3.0-next.4"
+    "vscode-ws-jsonrpc": "~3.3.0-next.5"
   },
   "devDependencies": {
     "monaco-editor-ms": "npm:monaco-editor@~0.47.0"
   },
   "peerDependencies": {
-    "@codingame/monaco-vscode-api": "~3.1.1",
-    "@codingame/monaco-vscode-editor-api": "~3.1.1",
-    "monaco-languageclient": "~8.0.0-next.4",
-    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.1"
+    "monaco-editor": "npm:@codingame/monaco-vscode-editor-api@~3.1.2",
+    "monaco-languageclient": "~8.0.0-next.5",
+    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.2"
   },
   "peerDependenciesMeta": {
-    "@codingame/monaco-vscode-api": {
-      "optional": false
-    },
-    "@codingame/monaco-vscode-editor-api": {
+    "monaco-editor": {
       "optional": false
     },
     "monaco-languageclient": {
