+++ verify/angular/package.json
@@ -18,11 +18,10 @@
     "@angular/core": "~17.2.4",
     "@angular/platform-browser": "~17.2.4",
     "@angular/platform-browser-dynamic": "~17.2.4",
-    "@codingame/monaco-vscode-api": "~3.1.1",
-    "@codingame/monaco-vscode-editor-api": "~3.1.1",
+    "monaco-editor": "npm:@codingame/monaco-vscode-editor-api@~3.1.2",
     "monaco-editor-wrapper": "../../packages/wrapper",
     "monaco-languageclient-examples": "../../packages/examples",
-    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.1",
+    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.2",
     "rxjs": "~7.8.1",
     "zone.js": "~0.14.4"
   },
