+++ verify/pnpm/package.json
@@ -4,11 +4,10 @@
   "private": true,
   "type": "module",
   "dependencies": {
-    "@codingame/monaco-vscode-api": "~3.1.1",
-    "@codingame/monaco-vscode-editor-api": "~3.1.1",
+    "monaco-editor": "npm:@codingame/monaco-vscode-editor-api@~3.1.2",
     "monaco-editor-wrapper": "../../packages/wrapper",
     "monaco-languageclient-examples": "../../packages/examples",
-    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.1"
+    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.2"
   },
   "devDependencies": {
     "http-server": "~14.1.1",
