+++ verify/yarn/package.json
@@ -4,13 +4,12 @@
   "private": true,
   "type": "module",
   "dependencies": {
-    "@codingame/monaco-vscode-api": "~3.1.1",
-    "@codingame/monaco-vscode-editor-api": "~3.1.1",
-    "@typefox/monaco-editor-react": "3.0.0-next.4",
-    "monaco-editor-wrapper": "4.0.0-next.4",
-    "monaco-languageclient-examples": "8.0.0-next.4",
-    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.1",
-    "vscode-ws-jsonrpc": "~3.3.0-next.4"
+    "@typefox/monaco-editor-react": "3.0.0-next.5",
+    "monaco-editor": "npm:@codingame/monaco-vscode-editor-api@~3.1.2",
+    "monaco-editor-wrapper": "4.0.0-next.5",
+    "monaco-languageclient-examples": "8.0.0-next.5",
+    "vscode": "npm:@codingame/monaco-vscode-api@~3.1.2",
+    "vscode-ws-jsonrpc": "~3.3.0-next.5"
   },
   "devDependencies": {
     "http-server": "~14.1.1",
